remove schema template glob -> regex cache
